Tests and score board but it does not work
scoreBoardDisplay.fxml will not display
tempdisplayScoreBoard in BlackBoxApplication is being used to display this. If the .fxml file is replaced with menu.fxml it works. I have checked scoreBoardDisplay.fxml in a dummy file and it worked so I do not know why it does not here. Please check if it works on your ide